# EventHub-BE/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "EventHub BE",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5678",
    "basePath": "/",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "login need token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credential",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Account"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Invalidate user bearer token for logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logout"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "logout success",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "401": {
                        "description": "invalid bearer token",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
        "/events/my": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get events created by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "My Events"
                ],
                "summary": "Get my events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
        "/notification/my": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get list of notifications for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get user notifications",
                "responses": {
                    "200": {
                        "description": "my notif retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.MyNotif"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "please login first / invalid token",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
        "/user/{eventId}/join": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Join an event by providing the event ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "join events"
                ],
                "summary": "Join an event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "joined event success",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "400": {
                        "description": "invalid event id",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "401": {
                        "description": "please login first / invalid token",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "409": {
                        "description": "you already joined this event",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "failed to join event",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
        "/user/{eventId}/leave": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Leave an event that the user previously joined",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leave events"
                ],
                "summary": "Leave an event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "left event successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "400": {
                        "description": "invalid event id",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "401": {
                        "description": "please login first / invalid token",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "404": {
                        "description": "you have not joined this event",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "failed to leave event",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Account": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "dto.Res": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "token": {}
            }
        },
...
        "model.MyNotif": {
            "type": "object",
            "required": [
                "id_notif"
            ],
            "properties": {
                "desk": {
                    "type": "string"
                },
                "id_notif": {
                    "type": "integer",
                    "minimum": 1
                },
                "read_at": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}